refactor: useModal hook 수정
[#41]
