+++ packages/frontend/src/hooks/useModal.ts
@@ -1,24 +1,14 @@
-import { useEffect } from "react";
+import { useState } from "react";
 
-import { ESC_KEY_CODE } from "../constants";
-
-export default function useModal({ onClose }: { onClose: () => void }) {
-  const escKeyCloseEvent = (event: KeyboardEvent) => {
-    if (event.key === ESC_KEY_CODE) {
-      onClose();
-    }
+export default function useModal() {
+  const [modalOpen, setModalOpen] = useState(false);
+  const closeModal = () => {
+    setModalOpen(false);
   };
 
-  useEffect(() => {
-    setScrollLock(true);
-    window.addEventListener("keydown", escKeyCloseEvent);
-    return () => {
-      window.removeEventListener("keydown", escKeyCloseEvent);
-      setScrollLock(false);
-    };
-  }, []);
-
-  const setScrollLock = (isLock: boolean) => {
-    document.body.style.overflow = isLock ? "hidden" : "auto";
+  const openModal = () => {
+    setModalOpen(true);
   };
+
+  return { modalOpen, openModal, closeModal };
 }
